msm: kgsl: Remove excessive log in kgsl_snapshot_get_object
kgsl_snapshot_get_object prints error log when gpu buffer is no
longer available in shared mem. When a list of buffers are not
available in shared memory, this log may lead to hog the cpu and
may lead to watchdog timeout.

Removed this error log as it was not printing any useful info.

Change-Id: I63eb58b9947fcaa4808334aa64806ea69c691dc6
Signed-off-by: Pankaj Gupta <gpankaj@codeaurora.org>
Signed-off-by: Harshitha Sai Neelati <hsaine@codeaurora.org>
Signed-off-by: feravolt <fera.volt@gmail.com>